AGSL: Add basic grammar
This CL adds a *very* basic grammar for AGSL. It does *not*
model the language, but applies some logical groupings to the
lexical tokens at the PSI level, and registers a ParserDefinition.
This allows for the language to be embedded in other languages,
such as having AGSL source inside Java/Kotlin strings.

Test: AgslParserTest
Bug: None
Change-Id: I7130c8ca19f2c35345497efd7fa0a3c5b0b582f9