Attempt to fix FastPreviewManagerTest flakiness
Write actions consumed on the event dispatch thread might invalidate a
file while the embedded compiler is doing work on it. Causing to
sometimes fail.

Write actions seem to be scheduled from the editor. So this change makes
sure Events are consumed before triggering the compiler, after making a
change.

Bug: 222319292
Test: reproduced locally with some tweaks
Change-Id: Ied49a921dba0115592f44c370713dcdb7f8cdd51