[AUA] Show information about other plugin upgrades
Compute and cache an initial set of usages for components, so that the
UI can display information about what it will do.  Update that cache
whenever the set of usages is recomputed.

Provide an `additionalInfo` hook for steps to provide pre-HTML-formatted
content (as opposed to the shortDescription which comes from the
processor), and use it in the AgpGradleVersionRefactoringProcessor's
step presentation.

Bug: 186391921
Test: added, manual
Change-Id: Ic582f20a9f3cf3128e149c6be1822d539e2d4539
(cherry picked from commit d275fee7e7ed92d83bbf748e3b7abe24d758d935)