Add groupByComposable option when sorting Compose Previews
With Multipreview it may happen that the Previews of a given
composable are defined in different places and even in different
files. This means that those Previews are no longer implicitly
grouped by their composable methods when sorting by sourceOffset.

This change adds an optional flag when sorting Previews that
allows its callers to specify whether an explicit group by
composable should be done or not. The deafult value for this
option will be determined by the Multipreview Studio Flag value.

Bug: 220889326
Bug: 220886387
Test: AnnotationFilePreviewElementFinderTest.testOrderingMultipreview
Change-Id: I306ba323c57b80d47bf431dca36a0ef9bc46e475