Allow a model to move instances in a tree move
Add the option: dndDeleteOriginOfInternalMove.
Normally the tranfer handler will delete the origin item on a
move. If the model prefer to move the instance instead of
creating a new instance for the target position, then the
transfer handler should not delete the item that was just moved.

Allow a client to specify false to dndDeleteOriginOfInternalMove
such that the internal moves can happen without a create/delete.

Bug: 210465107
Test: Unit tests updated/added
Change-Id: Idffba2651893353662a05b6a3b9f76ea6de1ea6d