[PreviewPicker] Add tracker implementation
Renamed PreviewPickerTracker -> ComposePickerTracker

ComposePickerTracker has a base implementation that handles the expected
usage of the interface.

Added a new PreviewPickerTracker with the relevant Preview
implementation details.

For now, the implementation just builds local data classes. These will
be changed once the studio_stats object is updated with the desired
classes/values on a follow-up change.

New PreviewPickerTracker implementation not in use yet.

Bug: 205184728
Test: PreviewPickerTrackerTest
Change-Id: I902149e77f08b7231833c904ac3eadc2d2ad4c29