Fix execution of Preview refresh requests to ensure once at a time
The behavior of collectLatest in the refresh flow was supposed to
cancel a running refresh Job when creating new ones, but it was
actually cancelling the wait on the Job, but not the Job itself.

This change fixes this by always collecting the latest RefreshRequest
and skipping the previous requests that were not yet processed. And
also making sure to wait for the running request before collecting
another one.

Fixes: 225058589
Test: manually by following the steps described in the bug.
Change-Id: Icc7de23c9968f612116c0d7da989b0100f4ab82d