[Compose-CL] Modify logic used to find ConstraintSets property
This is so that we can create the model for the ConstraintSets even when
the completion is invoked from an element outside the ConstraintSets
property, but within the same Json.

For example, when invoking completion within Transitions, we still want
the model for ConstraintSets even though Transitions is a sibling of
ConstraintSets.

Bug: 207030860
Test: passing, follow up CL will include test to cover this behavior
Change-Id: Ia3ce343d3edfc13e6917e7206bd6c6bf80918fef