LE: Retry on InvalidModuleException
Another analysis might invalidate the module we are currently
compiling during a long compile. In theory this can never be
the case given all analysis should be done with ReadActions.

There might be a bug somewhere that isn't part of LE or Studio.
For now, should we encounter such case, we cancel the current
compile and retry.

Test: NONE
Bug: 224875189
Change-Id: If5051a5c5ea4562117b5aed36d0006ed31ca7040