Handle Fast Preview changes faster
This replaces the previous setupChangeListener with a new
documentChangeFlow. This flow will generate an element per change. This
is used with debounce to handle changes with different timing depending
if Fast Preview is enabled (250ms debounce) or not (1s).

This also changes moves the runReadActionWithWritePriority to a lower
level so the process can be cancelled between retries. Before this
change, all the retries would be done before cancelling.

Bug: 199384558
Test: EmbeddedCompilerClientImplTest, ChangeManagerTest
Change-Id: I2f6502d41aa367389db86bd2256f0d898ba6dfb1