[PreviewPicker] Write a context dependent check for DeviceSpec Language
For DeviceSpec Language syntax, enforce that width,height,chinSize use
the same unit when declaring their value. So they should use either 'dp'
or 'px'.

A ParameterRule can now attempt to fix the value of the parameter, this
is so that user's input changes as little as possible if they've got
something wrong.

A ParameterRule may also now access arbitrary objects that persists
across all the value checks and fix attempts.

In this case, the rules for width,height and chinSize verify that they
all use the same unit and apply the necessary fix.

Bug: 220006785
Test: DimensionParameterRuleTest, PreviewAnnotationCheckTest
Change-Id: I3c5ab5888807ea98cdf5d6161e8f9343d13e5cab