[AUA] Introduce an error state into the AUA UX flow
This error state is used to display a diagnostic message (and disable
various graphical controls) when an internal error occurs; the
diagnostic message encourages the user to revert to a known-good state
before attempting to do anything else.

Introduce a new boolean to control whether the drop-down version combo
is active: it should be active in the previously-existing non-loading
states, but should be inactive if we have just hit an internal error.

Re-use the newly-introduced logic to disable the (empty) tree view of
Refactoring Processors in the AllDone state.

Bug: 227793240
Test: manual
Change-Id: Ifeef1f1c93f161a3f22e41a6bc8841e30d131a10