Change Compose ScreenView provider to use showSystemDecoration
This changes the behaviour from overriding the device configuration so
simply changing the behaviour of the ScreenView. Before this change,
depending of the showSystemDecoration value, the actual Device used in
the render Configuration would be changed from ROUND to SQUARE when
showSystemDecoration was false.

This caused that, in wear previews, the running code would receive the
wrong configuration based on a display setting.
This change ensures the Configuration device is constant and the
ScreenView devices on the shape to be used based on the value of
showSystemDecoration.

Fixes: 224727944
Test: ComposeScreenViewProvidersTest
Change-Id: Iaf488f56fa2a3daf29a59514b0939e8ec67d9c36