[BA] setup notification as regular with toolwindow balloon as default
Change implementation of notification to a more standard way.
Register it in plugin.xml, this will add the corresponding item
to settings. Set default to tool window balloon.
It also adds notification to the event log.
User can change it to be normal notification balloon or hide it
completely.

Remove previous notification setting from BA settings page, hide the
page as it currently ends up empty.
Migration was added for the case when user already have turned the
notification off, set it to NONE in the new system then.

Bug: 228294330
Bug: 259468778
Fixes: 260842548
Test: Adjusted existing
Change-Id: I65c7163d417868b51cf34ea66e3100570e3128d4