Take advantage of delayes parameter extraction
Send "delay_parameter_extractions" with the update settings command.
If the compose inspector agent understands this option, we will
start to receive an anchorHash for each ComposeViewNode.
(and an anchor hash is also included in references for drilling into
 parameter details.)

We will send the anchor hash in the commands for getting parameters
and parameter details.

Test that the:
- "delay_parameter_extractions" option is specified
- the anchor hash is specified in get parameter commands
- the anchor hash is specified in references
- getComposables has `extractAllParameters` set correctly

Fixes: 229657989
Test: Update existing tests
Change-Id: I58113b53a4179301f1f41a52dc24f3947265a53a