Make video codec configurable via StudioFlags
The codec name is passed to the agent by Studio, but the agent is free to use a different codec if it so chooses. The name of the codec chosen by the agent is included in the header of the video stream.

Add logging of video frames that are produced during the first second after a touch event to get a sense of codec responsiveness.

Experiments showed that VP8 is the most responsive during app start and close animations. VP9 is almost as responsive as VP8 on Pixel 3a API 30, but is much slower on Pixel 4 XL API T. H.264 is noticeably less responsive than VP8 on both, Pixel 3a API 30, and Pixel 4 XL API T.

Test: later
Bug: 229261814
Change-Id: I6f3b0acbf7f5dd4551ae648a0ddff7a47721bb65