Only sync when deploying if the ABI is avaliable
When deploying an app with native components, Studio would like to have
the debug information for the ABI of the deployment devices

With V2 native sync (since AGP 4.2) these are always avaliable
and are directly read from files. With V1 native sync these
might not be avaliable, and Studio currently triggers a sync
during deployment in those cases.

This change reduces the cases where sync might be triggered on every
deployment, but not actually fetch any new information.

Bug: 226372843
Test: Fixed and strengthened assertions in MakeBeforeRunTaskProviderTest
Change-Id: Ieffbbf879a6aeb35c8f8576c861502ac3bdd500d