Update ProcessNameMonitor to use child scopes
To make the code compatible with runBlockingTest and prod,
ProcessNameMonitor and ProcessNameClientMonitor now
both accept a parent CoroutineScope parameter in their
constructor. They both create a child scope that inherits
the context of the parent scope (dispatcher, etc.) and that
is also tied to their Disposable implementation so they are
cancelled then disposed.

Test: Existing
Bug: n/a
Change-Id: I01157c2bca79afe5853e0e34a066ba0916864469