Always complete the lint annotating process
The function LintHighlightingPass.collectInformation() is called in a
job thread, that thread may be canceled when platform stop collecting
backgrond highlighting data.

However, we want to make sure the lint annotating process can be done
in any case. Start another background task to run the process to avoid
the cancelation from platform.

Bug: 227392967
Test: manually.
Change-Id: Ic6a59dde092df7b70e88ad7904c4865cb14482b0