[Compose-CL] Make completion case-sensitive
If the LookupElements are not case sensitive, the options may show if
the completions don't match the user's expression and when completed,
the resulting expression will keep the incorrect casing.

This makes it so that completion is only shown when the original
expression matches the casing, so that the resulting expression will
always have the correct syntax.

Bug: 232011179
Bug: 207030860
Test: completionIsCaseSensitive
Change-Id: I79d32e74ceb65d50537dc692af48f3eb937bf5e8