[AUA] Make the AGP version processor less special
Now that we have the capacity for any component processor to block an
upgrade, the AGP version processor needs less special handling.  For
the most part, it can be treated as a component, though one which has
a particular importance in the user experience of the AGP Upgrade
Assistant tool and related dialogs.

Bug: 225529521
Bug: 231690925
Test: existing tests adjusted
Change-Id: I21dae6b2aa2872386a4dab16a309cb94ff0f7f83