[SDK Index] Use severity when reporting issues
There is a change in the SDK Index proto to now include a severity field
in the version labes. If used, this severity can be BLOCKING or
NON_BLOCKING and the meaning of it is wheter or not using this version
will block or not publishing. This new field defines the following types
of isses implicitely (sorted by severity):

1. Non complaint issues (always blocking, regardless of Severity field)
2. Critical issue (message from developer) with BLOCKING severity
3. Outdated library with BLOCKING severity
4. Outdated library with NON_BLOCKING severity (after some time it will
   be blocking)
5. Critical issue with NON_BLOCKING severity (will stay as NON_BLOCKING
   unless authors change it)

This change adds two new lint inspections to match changes in ag/18322382
and generates the PsIssues for PSD accordding to the severity in the
previous list.

Bug: 224643959
Test: Manual
Change-Id: I08afcdf43cdcec0fc6d69aa72c8733068d9fe501