[Dsl,.gr] Support limited completion for android.kotlinOptions
Whatever else happens in this area, we will need to special-case
kotlinOptions as it is grafted on to the android extension in an even
more dynamic (hence, even less statically inspectable) way than other
things.  This only provides limited completion because in Groovy-based
Gradle projects the resulting resolve scope of the KotlinJvmOptions
class is itself incomplete, so properties and methods from supertypes
are not found.

Bug: 155422585
Test: manual
Change-Id: I8a9c1db525f1f2b4e9e8dce4a1b60f4dad9d14ab