Do not use extension to detect source
The SourceCodeEditorProvider was using the extension to detect the file
type. This broken in Chipmunk and Dolphin because the VCS Diff preview
uses files that contain the extension. This caused the
SourceCodeEditorProvider to rely on the PreviewDiffVirtualFile which
does not have a Document and breaks.

This change forces the detection to use the VirtualFile FileType. The
FileType for Diff preview is a DiffFileType.

Fixes: 232045613
Test: SourceCodeEditorProviderTest and manually verified
Change-Id: Iee0b8d6445ecf1801a45f9c95f5442ffd89e8789
(cherry picked from commit ca73bd0cbd0e1a60f1e56b51908f5558a0067c2d)