Correctly record the number of Gradle projects
We used to use the number of IDE modules to record the number of Gradle
projects when tracking the project structure. This used to be roughly
correct since each Gradle project mapped to a single IDE modules.

As of switching on Module Per Source Set this number became wildly
wrong. This change re-adjusts the number recorded by traversing the
ExternalProject tree that the platform obtains from Gradle. Each
ExternalProject is built from exactly one Gradle project.

Fixes: 233115161
Test: test updated, mannually checked for some larger projects
Change-Id: If9c59c6ed3272a8b32c1eddf13924b775d36093f