Correctly load saved state when opening Image Asset
The bindTwoWay method of BindingsManager is not completely symmetrical
as when it is called it overrides the value of its first argument with
the value of the second.
This ensures that the icon generator values are the ones being overridden
when creating the binding, so that the state, which is loaded into the
fields of ConfigureAddaptiveIconPanel, is the one whose values have
priority.

Fixes: 228721461
Test: checked manually
Change-Id: I4a812937dc9934b5a284e9b0899eaf3a51d08fb4