Remove more dependence on project from StringResource.
This removes the pre-built String representing the StringResource's
directory. StringResource has no business knowing how UI will display
the directory name, and this causes it to be further dependent on what
Project it is in. Move that functionality out to an area where we already
have the Project information or can get it in a way that makes more sense.

Bug: 232444069
Test: Existing presubmits
Change-Id: I89edb828bf2a7192d84ddcb17123e18e790a7d21