Maintain the scroll position in the resource manager.
The lists in the resource manager are rebuilt every time there is
a change in its resources, and the scroll position is reset to the
top of the list each time. However, when a change occurs in a
resource that is part of the list that is being shown at the moment,
then the scroll position shouldn't be affected.

Fixes: 169796111
Test: manually verified
Change-Id: If174bac9f9d8e19325a6d5b25058b371660f65da