Change AndroidStudio close to kill on close
We must terminate the process on close. If we don't and
expect the test to gracefully terminate it always, it means
that if the test has an assertEquals, when the assertion
exception is thrown the try-catch will attempt to close
the studio object that has not been asked to terminate,
blocking forever until the test times out, swallowing the
assertion information.

Test: this is for tests
Bug: NA
Change-Id: I5574e8bc0371d0613227092d9c90b152d166d5d0