Load missing R ids from external libraries
When non-transitive R classes are enabled, the resource ids from
resources in external libraries are not added to the module R class.
Hence we must look for the R classes from external dependencies and
parse those as well as the project R classes.
When parsing R classes from external dependencies, we must also account
for whether the module we are considering is using final ids or not. If
it is we must make sure that we do not create a light R class version in
the LibraryResourceClassLoader, but instead we must let the class loader
find the compiled R classes.

Bug: 233862429
Test: ModuleClassLoaderTest, LibraryResourceClassLoaderTest
Change-Id: I7c536274e39a6028d0895abf61ff5d0a7bcabc5a
(cherry picked from commit ff8bf5180c88492d20bb2bfaaa1976de9300bef1)