Fixes process shutdown
Destroying the process for studio.sh does not give the
result we want, similarly to send a kill to it, it leaves
the IDE running and only the shell process is ended.

Additionally ::attach and ::run worked differently in the
sense that attach would indeed hold a reference to the IDE
process.

This change unifies both approaches and fixes the termination
issue.

Test: this is for tests
Bug: NA
Change-Id: I1928d04a70576bf198baee0f9b9787674de3be6a