[BA] Split result for disabled and not active downloads analyzer
Effectively there are three cases when there might be no download events
and two of them are special:
 1) analyzer is disabled completely with the flag,
 2) gradle version is old and does not provide any data,
 3) there are simply no downloads in this build.
Before this CL 1) and 2) were merged in single state but it is not handy
as we want to present them separately. e.g. in case of 1) there should
not be any trace of Downloads analysis on the UI and in case of 2) we
need to mention that gradle needs to be upgraded.

Bug: 218505954
Test: Updated existing tests
Change-Id: I27aa69d306c93837c9cc83da90a89937dad35e30