Improve handhsake
If a device is connected to Studio while locked, the current handshake
implementation does not work. As the top activity wouldn't be found.
Therefore the handshake would terminate saying that the device does
not support foreground process detection, which might be a false
negative.

To work around this limitation, the handshake has been changed.
Now the handshake has three possible result states:

1. SUPPORTED
2. NOT_SUPPORTED
3. UNKNOWNW

This change handles these new results of the handshake process.

SUPPORTED and NOT_SUPPORTED are self explicative.
UNKNOWN means that at the moment we can't determine if foreground
process detection is supported or not.

When we get an UNKNOWN resposne, we should repeate the handshake
periodically, until it results in SUPPORTED or NOT_SUPPORTED.

Test: ForegroundProcessDetectionTest.kt
Bug: 232067042
Change-Id: If4979d94f9e61edc0b4cdfce093d8ee00f67add2