Fix R classes not being loaded before project is built
We've recently made a change to address an issue of R classes not being
loaded in compiled code. Basically, we've moved the class loading of R
classes from LibraryResourceClassLoader, whose loading mechanism relies
on generated light classes, to ProjectSystemClassLoader, which loads
classes from compiled code.

The problem is we overlooked the scenario of users being able to see XML
layouts before the project is built, exactly because the R classes are
loaded from load classes. After our change, this was not possible, since
we'd try to load the R classes from compiled code, which does not exist
before a build (or after a clean).

This change fixes this issue by using LibraryResourceClassLoader as a
fallback, in case the project has not been built, i.e. compiled classes
are not available.

Bug: 236854148
Bug: 233862429
Bug: 236141674
Test: Regression test added to LibraryResourceClassLoaderTest + UI tests
Change-Id: Ief55af48824318f7ac8ee3954e05ba5c85448d0b
(cherry picked from commit 65625eba0902d632f8e742ab9eec1fd7c533c06d)