[Dsl,.toml] Support creation of version.ref elements
We need to intercede in the creation of children of map elements in
the `libraries` table.  As we represent `libraries` as an ExtModel, we
need a hook so that children of this model are distinguished;
introduce factory parameters that default to GradlePropertyModelImpl
creators, but can be overridden.

Define a subclass of GradlePropertyModelImpl whose only difference
manifests when getting a map value from the property with a key of
"version".  In that case, construct a property with a transform, which
creates a GradleDslVersionLiteral (rather than a regular
GradleDslLiteral) if no element is present.

Bug: 238166819
Test: added
Change-Id: I6d430d37bd4eaa029ac163a0541f5257058adefc